✅refactor(LadderGameController): Controller를 하나로 합체
- 분리 되어있던 로직을 하나로 합체
- Enum 과 ResultFinder 분리에 의한 수정
- Return List가 아닌 Map으로 변경
- Application, Controller 통합에 따른 수정
