+++ src/main/java/ladder/Application.java
@@ -1,7 +1,6 @@
 package ladder;
 
 import ladder.controller.LadderGameController;
-import ladder.controller.LadderResultController;
 import ladder.domain.randomGenerator.RandomEnergyGenerator;
 import ladder.view.InputView;
 import ladder.view.OutputView;
@@ -14,7 +13,6 @@ public class Application {
         LadderGameController ladderGameController = new LadderGameController(new RandomEnergyGenerator(), inputView,
                 outputView);
 
-        LadderResultController ladderResultController = ladderGameController.run();
-        ladderResultController.showResult();
+        ladderGameController.run();
     }
 }
